requirements.txt
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.5.0
numpy>=1.21.0
pytest>=7.0.0
PyYAML>=6.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
